pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.4",
    "pynacl>=1.6.1",
    "pytz>=2025.2",
    "yt-dlp>=2025.10.22",
]
